pyproject.toml
[project]
name = "chatterbox-ui"
version = "0.0.1"
description = "ChatterboxUI: A Simple UI Wrapper for Chatterbox-TTS!"
readme = "README.md"
requires-python = ">=3.10,<3.12" # newest numpy version allowed requires distutils, which is dropped in 3.12
license = {file = "LICENSE"}
authors = [
    {name = "Lopolin"}
]
dependencies = [
    "chatterbox-tts>=0.1.4",
    "gradio",
    "hf_xet",
    "numpy",
    "ml-dtypes>=0.5.0",
    "pandas>=2.2.0", # This introduced future.no_silent_downcasting; chatterbox-tts 1.2 used 2.3.3 of pandas, but i hope this is good enough
    "pipdeptree>=2.29.0", # for checking the dep tree since uv tree is not expressive enough
]

# Fix some build error; recommended by UV itself, don't question the preview warning
[tool.uv.extra-build-dependencies]
pkuseg = ["numpy"]

# [project.urls]
# Homepage = "https://github.com/resemble-ai/chatterbox"
# Repository = "https://github.com/resemble-ai/chatterbox"

# [build-system]
# requires = ["setuptools>=61.0"]
# build-backend = "setuptools.build_meta"

# [tool.setuptools.packages.find]
# where = ["src"]
